Add container owner & self user check

+++ jiacpp-platform/src/main/java/de/dailab/jiacpp/platform/UserController.java
@@ -1,14 +1,20 @@
 package de.dailab.jiacpp.platform;
 
 import de.dailab.jiacpp.model.User;
+import de.dailab.jiacpp.platform.auth.JwtUtil;
 import de.dailab.jiacpp.platform.user.TokenUserDetailsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.extern.java.Log;
+import org.apache.hc.client5.http.HttpResponseException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -21,6 +27,9 @@ public class UserController {
 
     @Autowired
     TokenUserDetailsService userDetailsService;
+
+    @Autowired
+    private JwtUtil jwtUtil;
 
     /*
      * USER MANAGEMENT
@@ -42,6 +51,7 @@ public class UserController {
     @RequestMapping(value="/users/{username}", method=RequestMethod.DELETE)
     @Operation(summary="Delete an existing user from the connected database", tags={"users"})
     public boolean deleteUser(@PathVariable String username) {
+        if (!isAdminOrSelf(username)) throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         log.info(String.format("DELETE USER: %s", username));
         return userDetailsService.removeUser(username);
     }
@@ -49,6 +59,7 @@ public class UserController {
     @RequestMapping(value="/users/{username}", method=RequestMethod.GET)
     @Operation(summary="Get an existing user from the connected database", tags={"users"})
     public String getUser(@PathVariable String username) {
+        if (!isAdminOrSelf(username)) throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         log.info(String.format("GET USER: %s", username));
         return userDetailsService.getUser(username);
     }
@@ -65,6 +76,7 @@ public class UserController {
     public String updateUser(
             @PathVariable String username,
             @RequestBody User user) {
+        if (!isAdminOrSelf(username)) throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         String logOut = String.format("UPDATE USER: %s (", username);
         if (user.getUsername() != null) logOut += String.format("NEW USERNAME: %s ", user.getUsername());
         if (user.getPassword() != null) logOut += "NEW PASSWORD ";
@@ -80,5 +92,19 @@ public class UserController {
             userRoles.put(role.getName(), role.getPrivileges());
         }
         return userRoles;
+    }
+
+    // Helper methods
+
+    /**
+     * Checks if the current request user is either an admin (has full control over user management)
+     * or the request user is performing request on its own data
+     * @param username: Name of user which will get affected by request (NOT THE CURRENT REQUEST USER)
+     * */
+    private boolean isAdminOrSelf(String username) {
+        UserDetails details = userDetailsService.loadUserByUsername(jwtUtil.getCurrentRequestUser());
+        if (details == null) return false;
+        return details.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN")) ||
+                details.getUsername().equals(username);
     }
 }
